+++ src/php/_theme.php
@@ -20,18 +20,16 @@
 
 // --- setup I18N --------------------------------------------------------------
 
-require_once('core/Translate.php');
-at\nerdreich\Translate::loadLanguage(dirname(__FILE__) . '/I18N/' . $config['language'] . '.yaml');
+require_once(dirname(__FILE__) . '/../../core/Translate.php');
+at\nerdreich\wiki\Translate::loadLanguage(dirname(__FILE__) . '/I18N/' . $wiki->getLanguage() . '.yaml');
 
 // --- register theme macros ---------------------------------------------------
 
-$wiki->registerMacro('pagelist', function (
+$wiki->core->getPlugin('macro')->registerMacro('pagelist', function (
     ?string $primary,
     ?array $secondary,
     string $path
-): string {
-    global $wiki;
-
+) use ($wiki): string {
     $dir = dirname($path);
     $snippet = '';
     $cols = [];
@@ -55,13 +53,13 @@ $wiki->registerMacro('pagelist', function (
     foreach (scandir($dir) as $filename) {
         if (preg_match($pattern, $filename)) {
             $absFilename = $dir . '/' . $filename;
-            list($metadata, $content) = $wiki->loadFile($absFilename, true);
+            list($metadata, $content) = $wiki->core->loadFile($absFilename, true);
             if ($primary === '#grouplist') {
                 $snippet .= '[' . $metadata['title'] . '](' .
-                    $wiki->findURLPathForContentFile($absFilename) . ')  ' . PHP_EOL;
+                    $wiki->core->contentFileFSToWikiPath($absFilename) . ')  ' . PHP_EOL;
             } else {
                 $snippet .= '|[' . $metadata['title'] . '](' .
-                    $wiki->findURLPathForContentFile($absFilename) . ')|';
+                    $wiki->core->contentFileFSToWikiPath($absFilename) . ')|';
                 foreach ($cols as $col) {
                     preg_match_all('/^' . $col . ':\s+(.*)/m', $content, $field);
                     $snippet .= $field[1][0] . ' |';
@@ -74,7 +72,7 @@ $wiki->registerMacro('pagelist', function (
     return $snippet;
 });
 
-$wiki->registerMacro('paginate', function (
+$wiki->core->getPlugin('macro')->registerMacro('paginate', function (
     ?string $primary,
     ?array $secondary,
     string $path
@@ -111,37 +109,25 @@ $wiki->registerMacro('paginate', function (
     return $snippet;
 });
 
-// --- HTML snippets -----------------------------------------------------------
+// --- register filters --------------------------------------------------------
 
-function getPageLinksHTML(
-    $user,
-    $wiki
-): string {
-    $html = '';
-    if ($user->mayUpdate($wiki->getWikiPath())) {
-        if ($wiki->exists()) {
-            $html .= '<a href="?action=edit"><i class="fas fa-edit"></i> ' . ___('Edit') . '</a><br>';
-        } else {
-            $html .= '<a href="?action=createPage"><i class="fas fa-edit"></i> ' . ___('Create') . '</a><br>';
-        }
-    }
-    if ($wiki->exists() && $user->mayRead($wiki->getWikiPath()) && $user->mayUpdate($wiki->getWikiPath())) {
-        $html .= '<a href="?action=history"><i class="fas fa-history"></i> ' . ___('History') . '</a><br>';
-    }
-    if ($wiki->exists() && $user->mayDelete($wiki->getWikiPath())) {
-        $html .= '<a href="?action=delete"><i class="fas fa-trash"></i> ' . ___('Delete') . '</a><br>';
-    }
-    if ($user->mayAdmin($wiki->getWikiPath())) {
-        $html .= '<a href="./?admin=folder"><i class="fas fa-folder"></i> ' . ___('Permissions') . '</a><br>';
-    }
-    if ($user->isLoggedIn()) {
-        $html .= '<p>' . $_SESSION['username'] . '</p>';
-        $html .= '<a href="?auth=logout"><i class="fas fa-sign-out-alt"></i> ' . ___('Logout') . '</a>';
-    } else {
-        $html .= '<a href="?auth=login"><i class="fas fa-sign-in-alt"></i> ' . ___('Login') . '</a>';
-    }
+$wiki->core->registerFilter('html', 'fontawesome', function (string $html, string $path): string {
+    // fontawesome instead of certain emojis
+    $html = str_replace('👤', '<i class="fas fa-user"></i>', $html);
+    $html = str_replace('👥', '<i class="fas fa-users"></i>', $html);
+    $html = str_replace('🎬', '<i class="fas fa-film"></i>', $html);
+    $html = str_replace('📖', '<i class="fas fa-theater-masks"></i>', $html);
+    $html = str_replace('♖', '<i class="fas fa-chess-knight"></i>', $html);
+    $html = str_replace('📅', '<i class="far fa-calendar-alt"></i>', $html);
+    $html = str_replace('🗄️', '<i class="fas fa-archive"></i>', $html);
+    $html = str_replace('💬', '<i class="fas fa-theater-masks"></i>', $html);
+    $html = str_replace('🏠', '<i class="fas fa-home"></i>', $html);
+    $html = str_replace('🗺', '<i class="fas fa-map-marked-alt"></i>', $html);
+    $html = str_replace('🪶', '<i class="fas fa-feather-alt"></i>', $html);
     return $html;
-}
+});
+
+// --- other theme helpers -----------------------------------------------------
 
 /**
  * Convert a wiki path into a series of CSS classes.
@@ -200,49 +186,85 @@ function diff2html(
 }
 
 /**
+ * Convert a date or time to a string based on the current language/locale.
+ *
+ * Will auto-detect the format of the date.
+ *
+ * @param mixed $param A date of some kind.
+ * @return string Formatted date.
+ */
+function localDateString(
+    $param
+): string {
+    global $wiki;
+    if (gettype($param) === 'object' && get_class($param) === 'DateTime') {
+        return $param->format($wiki->getDateTimeFormat());
+    }
+    if (gettype($param) === 'integer') {
+        return (new \DateTime('@' . $param))->format($wiki->getDateTimeFormat());
+    }
+    return $param;
+}
+
+// --- output ------------------------------------------------------------------
+
+/**
+ * Assemble the navigation menu.
+ *
+ * @param\at\nerdreich\wiki\WikiUI $wiki Current UI object.
+ */
+function getPageLinksHTML(at\nerdreich\wiki\WikiUI $wiki): string
+{
+    $html = '';
+    foreach ($wiki->getMenuItems() as $action => $label) {
+        $html .= '<a href="?' . $action . '">' . ___($label) . '</a><br>';
+    }
+    return $html;
+}
+
+/**
  * Generate the HTML header and open the <body>.
  *
- * @param at\nerdreich\Wiki $wiki Current CMS object.
- * @param array $config Wiki configuration.
+ * @param\at\nerdreich\wiki\WikiUI $wiki Current UI object.
  */
-function outputHeader(array $config, string $path, string $title, string $description = '')
+function outputHeader(at\nerdreich\wiki\WikiUI $wiki, ?string $title = null, ?string $description = null): void
 {
     ?><!doctype html>
 <html class="no-js" lang="">
 <head>
   <meta charset="utf-8">
-  <title><?php echo htmlspecialchars($title); ?></title>
-  <meta name="description" content="<?php echo htmlspecialchars($description); ?>">
+    <?php
+    echo $title === null ? '' : '<title>' . htmlspecialchars($title) . '</title>';
+    echo $description === null ? '' : '<meta name="description" content="' . htmlspecialchars($description) . '">';
+    ?>
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="manifest" href="<?php echo $config['themePath']; ?>site.webmanifest">
-  <link rel="apple-touch-icon" href="<?php echo $config['themePath']; ?>icon.png">
-  <link rel="icon" href="<?php echo $config['themePath']; ?>favicon.ico"  type="image/x-icon">
-  <style>body { background-color: $BGCOLOR$; color: rgba(255, 255, 255, 0.8); }</style>
-  <link rel="stylesheet" href="<?php echo $config['themePath']; ?>style.css?v=$VERSION$">
+  <link rel="manifest" href="<?php echo htmlspecialchars($wiki->getThemePath()); ?>site.webmanifest">
+  <link rel="apple-touch-icon" href="<?php echo htmlspecialchars($wiki->getThemePath()); ?>icon.png">
+  <link rel="icon" href="<?php echo htmlspecialchars($wiki->getThemePath()); ?>favicon.ico"  type="image/x-icon">
+  <link rel="stylesheet" href="<?php echo htmlspecialchars($wiki->getThemePath()); ?>style.css?v=$VERSION$">
 </head>
-<body class="<?php echo htmlspecialchars(pathToClasses($path)); ?>">
+<body class="<?php echo htmlspecialchars(pathToClasses($wiki->core->getWikiPath())); ?>">
     <?php
 }
 
 /**
  * Generate the (top) navbar.
  *
- * @param at\nerdreich\Wiki $wiki Current CMS object.
- * @param at\nerdreich\UserSession $user Current user/Session object.
+ * @param\at\nerdreich\wiki\WikiUI $wiki Current UI object.
  */
-function outputNavbar(at\nerdreich\Wiki $wiki, at\nerdreich\UserSession $user)
+function outputNavbar(at\nerdreich\wiki\WikiUI $wiki): void
 {
     ?>
 <section class="navbar">
   <nav class="container">
     <div class="row">
       <div class="col-12">
-        <?php echo beautify($wiki->getSnippetHTML('topnav')); ?>
+        <?php echo $wiki->core->getSnippetHTML('topnav'); ?>
         <div>
           <input id="wiki-burger" type="checkbox">
-          <label for="wiki-burger"><i class="fas fa-cog"></i></label>
+          <label for="wiki-burger"><svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="#ffffff" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><line x1="3" y1="12" x2="21" y2="12"></line><line x1="3" y1="6" x2="21" y2="6"></line><line x1="3" y1="18" x2="21" y2="18"></line></svg></label>
           <div class="wiki-menu">
-            <?php echo getPageLinksHTML($user, $wiki); ?>
+            <?php echo getPageLinksHTML($wiki); ?>
           </div>
         </div>
       </div>
@@ -255,16 +277,16 @@ function outputNavbar(at\nerdreich\Wiki $wiki, at\nerdreich\UserSession $user)
 /**
  * Generate the banner area.
  *
- * @param at\nerdreich\Wiki $wiki Current CMS object.
+ * @param\at\nerdreich\wiki\WikiUI $wiki Current UI object.
  */
-function outputBanner(at\nerdreich\Wiki $wiki)
+function outputBanner(at\nerdreich\wiki\WikiUI $wiki): void
 {
     ?>
 <section class="banner">
   <nav class="container">
     <div class="row">
       <div class="col-12">
-        <?php echo beautify($wiki->getSnippetHTML('banner')); ?>
+        <?php echo $wiki->core->getSnippetHTML('banner'); ?>
       </div>
     </div>
   </nav>
@@ -275,19 +297,19 @@ function outputBanner(at\nerdreich\Wiki $wiki)
 /**
  * Generate the footer and close <body> & <html>.
  *
- * @param at\nerdreich\Wiki $wiki Current CMS object.
+ * @param\at\nerdreich\wiki\WikiUI $wiki Current UI object.
  */
-function outputFooter(at\nerdreich\Wiki $wiki)
+function outputFooter(at\nerdreich\wiki\WikiUI $wiki): void
 {
     ?>
 <footer class="container">
   <div class="row">
     <div class="col-12">
       <p>
-        <a class="no-icon" href="<?php echo $wiki->getRepo(); ?>">wiki.md v<?php echo $wiki->getVersion(); ?></a>
+        <a class="no-icon" href="<?php echo $wiki->getRepo(); ?>">wiki.md v<?php echo $wiki->core->getVersion(); ?></a>
         - <a class="no-icon" href="$URL$">Sunset Theme v$VERSION$</a>
-        <?php if ($wiki->getDate() !== null) {
-            echo '- ' . htmlspecialchars(___('Last saved %s', $wiki->getDate()->format($config['datetime'])));
+        <?php if ($wiki->core->getDate() !== null) {
+            echo '- ' . htmlspecialchars(___('Last saved %s', localDateString($wiki->core->getDate())));
         } ?>
         - <a href="/<?php __('Privacy'); ?>"><?php __('Privacy'); ?></a>
       </p>
@@ -297,31 +319,4 @@ function outputFooter(at\nerdreich\Wiki $wiki)
 </body>
 </html>
     <?php
-}
-
-// --- other helpers -----------------------------------------------------------
-
-/**
- * Do various gentle content improvements, like icons ...
- *
- * @param string $html HTML to improve.
- * @return string Improved HTML.
- */
-function beautify(
-    string $html
-): string {
-    // fontawesome instead of emoji
-    $html = str_replace('👤', '<i class="fas fa-user"></i>', $html);
-    $html = str_replace('👥', '<i class="fas fa-users"></i>', $html);
-    $html = str_replace('🎬', '<i class="fas fa-film"></i>', $html);
-    $html = str_replace('📖', '<i class="fas fa-theater-masks"></i>', $html);
-    $html = str_replace('♖', '<i class="fas fa-chess-knight"></i>', $html);
-    $html = str_replace('📅', '<i class="far fa-calendar-alt"></i>', $html);
-    $html = str_replace('🗄️', '<i class="fas fa-archive"></i>', $html);
-    $html = str_replace('💬', '<i class="fas fa-theater-masks"></i>', $html);
-    $html = str_replace('🏠', '<i class="fas fa-home"></i>', $html);
-    $html = str_replace('🗺', '<i class="fas fa-map-marked-alt"></i>', $html);
-    $html = str_replace('🪶', '<i class="fas fa-feather-alt"></i>', $html);
-
-    return $html;
 }